Update modeSwitch() to clear the high bytes of X and Y in emulation mode.
Replace the interrupt system with a simple on/off status; the logic
for handling multiple interrupt sources has been pushed up to the
System class.

Fix RTI instruction to properly pull or not pull the PBR when the
CPU is switched between native and interrupt mode by the RTI
instruction.