This commit does three things:
1. reorganize the code (again). xgscore is now emulator, video is now vgc,
and most of the chip classes have been moved to their own folders.

2. Move the still relevant bits of docs/ out to the top level, and removes
the rest.

3. Move the actual video draw code up from VGC.cc into a new Video class
that is part of the emulator module. The VGC class now simply maintains a
flat frame buffer, and the emulator passes that frame buffer to the Video
class every tick to draw it. This change is in preparation for moving the
actual rendering over to OpenGL, which is needed for the GUI.