Make 5.25" disks bootable via slot 6 again. This commit fixes a lot of silly bugs in the IWM and Disk525 code.
A more impactful part of this commit is the move of the master CPU
cycle counter from System to Processor. The copy in System was only
updated after cpu->runUntil() was called, so it impossible to get
correct cycle counts at the instruction level. Such timing is
required for the IWM emulation.